+++ src/RingGrids/interpolation.jl
@@ -301,8 +301,8 @@ end
 
 function update_locator!(
     I::AbstractInterpolator{NF, Grid},  # GridGeometry and Locator
-    λs::Vector,                         # longitudes to interpolate onto
-    θs::Vector;                         # latitudes to interpolate onto
+    λs::AbstractVector,                 # longitudes to interpolate onto
+    θs::AbstractVector;                 # latitudes to interpolate onto
     unsafe::Bool=false,                 # true to disable safety checks
     ) where {NF<:AbstractFloat, Grid<:AbstractGrid}
 
@@ -315,11 +315,11 @@ function update_locator!(
     find_grid_indices!(I, λs)               # next points left and right of λ on rings north and south
 end
 
-function find_rings!(   js::Vector{<:Integer},  # Out: ring indices j
-                        Δys::Vector,            # Out: distance fractions to ring further south
-                        θs::Vector,             # latitudes to interpolate onto
-                        latd::Vector;           # latitudes of the rings on the original grid
-                        unsafe::Bool=false)     # skip safety checks when true
+function find_rings!(   js::AbstractVector{<:Integer},  # Out: ring indices j
+                        Δys::AbstractVector,            # Out: distance fractions to ring further south
+                        θs::AbstractVector,             # latitudes to interpolate onto
+                        latd::AbstractVector;           # latitudes of the rings on the original grid
+                        unsafe::Bool=false)             # skip safety checks when true
     
     if ~unsafe
         θmin, θmax = extrema(θs)
@@ -338,10 +338,10 @@ function find_rings!(   js::Vector{<:Integer},  # Out: ring indices j
     find_rings_unsafe!(js, Δys, θs, latd)
 end
 
-function find_rings_unsafe!(js::Vector{<:Integer},  # Out: vector of ring indices
-                            Δys::Vector,            # distance fractions to ring further south
-                            θs::Vector,             # latitudes of points to interpolate onto
-                            latd::Vector{NF},       # latitudes of rings (90˚ to -90˚, strictly decreasing)
+function find_rings_unsafe!(js::AbstractVector{<:Integer},  # Out: vector of ring indices
+                            Δys::AbstractVector,            # distance fractions to ring further south
+                            θs::AbstractVector,             # latitudes of points to interpolate onto
+                            latd::AbstractVector{NF},       # latitudes of rings (90˚ to -90˚, strictly decreasing)
                             ) where {NF<:AbstractFloat}
 
     @boundscheck length(js) == length(θs) || throw(BoundsError)
@@ -365,15 +365,15 @@ function find_rings_unsafe!(js::Vector{<:Integer},  # Out: vector of ring indice
 end
 
 # for testing only
-function find_rings(θs::Vector, latd::Vector{NF}) where NF
+function find_rings(θs::AbstractVector, latd::AbstractVector{NF}) where NF
     js = Vector{Int}(undef, length(θs))
     Δys = Vector{NF}(undef, length(θs))
     find_rings!(js, Δys, θs, latd)
     return js, Δys
 end
 
-function find_grid_indices!(I::AnvilInterpolator,   # update indices arrays
-                            λs::Vector)             # based on new longitudes λ
+function find_grid_indices!(I::AnvilInterpolator,       # update indices arrays
+                            λs::AbstractVector)         # based on new longitudes λ
 
     (; js, ij_as, ij_bs, ij_cs, ij_ds ) = I.locator
     (; Δabs, Δcds ) = I.locator
@@ -432,10 +432,10 @@ $(TYPEDSIGNATURES)
 Computes the average at the North and South pole from a given grid `A` and it's precomputed
 ring indices `rings`. The North pole average is an equally weighted average of all grid points
 on the northern-most ring. Similar for the South pole."""
-function average_on_poles(  A::AbstractVector{NF},
-                            rings::Vector{<:UnitRange{<:Integer}}
-                            ) where {NF<:AbstractFloat}
-    
+function average_on_poles(  
+    A::AbstractVector{NF},
+    rings::AbstractVector{<:UnitRange{<:Integer}},
+) where {NF<:AbstractFloat}    
     A_northpole = mean(view(A, rings[1]))    # average of all grid points around the north pole
     A_southpole = mean(view(A, rings[end]))  # same for south pole
     return convert(NF, A_northpole), convert(NF, A_southpole)
@@ -445,10 +445,10 @@ end
 $(TYPEDSIGNATURES)
 Method for `A::Abstract{T<:Integer}` which rounds the averaged values
 to return the same number format `NF`."""
-function average_on_poles(  A::AbstractGrid{NF},
-                            rings::Vector{<:UnitRange{<:Integer}}
-                            ) where {NF<:Integer}
-    
+function average_on_poles(  
+    A::AbstractGrid{NF},
+    rings::AbstractVector{<:UnitRange{<:Integer}}
+) where {NF<:Integer}
     A_northpole = mean(view(A, rings[1]))    # average of all grid points around the north pole
     A_southpole = mean(view(A, rings[end]))  # same for south pole
     return round(NF, A_northpole), round(NF, A_southpole)