+++ src/SpeedyWeather.jl
@@ -139,8 +139,8 @@ include("physics/surface_fluxes.jl")
 include("physics/convection.jl")
 include("physics/zenith.jl")
 include("physics/optical_depth.jl")
-include("physics/shortwave_radiation.jl")
 include("physics/longwave_radiation.jl")
+include("physics/shortwave_radiation.jl")
 include("physics/stochastic_physics.jl")
 
 # OCEAN AND LAND