+++ src/output/variables/surface_fluxes.jl
@@ -3,7 +3,7 @@ Fields are $(TYPEDFIELDS)"""
 @kwdef mutable struct SensibleHeatFluxOutput <: AbstractOutputVariable
     name::String = "sensible_heat_flux"
     unit::String = "W/m^2"
-    long_name::String = "Sensible heat fluxes (positive down)"
+    long_name::String = "Sensible heat fluxes (positive up)"
     dims_xyzt::NTuple{4, Bool} = (true, true, false, true)
     missing_value::Float64 = NaN
     compression_level::Int = 3
@@ -12,14 +12,14 @@ Fields are $(TYPEDFIELDS)"""
 end
 
 path(::SensibleHeatFluxOutput, simulation) =
-    simulation.diagnostic_variables.physics.surface_flux_heat
+    simulation.diagnostic_variables.physics.sensible_heat_flux
 
 """Defines netCDF output for a specific variables, see `VorticityOutput` for details.
 Fields are $(TYPEDFIELDS)"""
 @kwdef mutable struct EvaporativeFluxOutput <: AbstractOutputVariable
     name::String = "evaporative_flux"
     unit::String = "kg/s/m^2"
-    long_name::String = "Surface humidity fluxes (positive down)"
+    long_name::String = "Surface humidity fluxes (positive up)"
     dims_xyzt::NTuple{4, Bool} = (true, true, false, true)
     missing_value::Float64 = NaN
     compression_level::Int = 3
@@ -28,7 +28,7 @@ Fields are $(TYPEDFIELDS)"""
 end
 
 path(::EvaporativeFluxOutput, simulation) =
-    simulation.diagnostic_variables.physics.surface_flux_humid
+    simulation.diagnostic_variables.physics.evaporative_flux
 
 SurfaceFluxesOutput() = (
     SensibleHeatFluxOutput(),